+++ Coimbra.Services.Events/NullEvent.cs
@@ -1,0 +1,4 @@
+namespace Coimbra.Services.Events
+{
+    internal partial struct NullEvent : IEvent { }
+}